% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Accepts two numeric values, add them and return a numeric value.}
\usage{
add(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
A numeric value representing sum of x and y arguments to the function.
}
\description{
Accepts two numeric values, add them and return a numeric value.
}
\examples{
add(2,4)
add(78, 5)
}
